Conditional mutual information API

The condition mutual information (CMI) API is defined by

ConditionalMutualInformation, and its subtypes.

condmutualinfo,

ConditionalMutualInformationEstimator, and its subtypes.

CausalityTools.condmutualinfo — Function

condmutualinfo([measure::CMI], est::CMIEstimator, x, y, z) → cmi::Real

Estimate a conditional mutual information (CMI) of some kind (specified by measure), between x and y, given z, using the given dedicated ConditionalMutualInformationEstimator, which may be discrete, continuous or mixed.

Estimators

Estimator	Principle	CMIShannon	CMIRenyiPoczos

FPVP	Nearest neighbors	✓	x

MesnerShalisi	Nearest neighbors	✓	x

Rahimzamani	Nearest neighbors	✓	x

PoczosSchneiderCMI	Nearest neighbors	x	✓

condmutualinfo([measure::CMI], est::ProbabilitiesEstimator, x, y, z) → cmi::Real ∈ [0, a)

Estimate the conditional mutual information (CMI) measure between x and y given z using a sum of entropy terms, without any bias correction, using the provided ProbabilitiesEstimator est. If measure is not given, then the default is CMIShannon().

With a ProbabilitiesEstimator, the returned cmi is guaranteed to be non-negative.

Estimators

Estimator	Principle	CMIShannon	CMIRenyiSarbu

CountOccurrences	Frequencies	✓	✓

ValueHistogram	Binning (histogram)	✓	✓

SymbolicPermutation	Ordinal patterns	✓	✓

Dispersion	Dispersion patterns	✓	✓

condmutualinfo([measure::CMI], est::DifferentialEntropyEstimator, x, y, z) → cmi::Real

Estimate the mutual information between x and y conditioned on z, using the differential version of the given conditional mutual information (CMI) measure. The DifferentialEntropyEstimator est must must support multivariate data. No bias correction is performed. If measure is not given, then the default is CMIShannon().

Note

DifferentialEntropyEstimators have their own base field which is not used here. Instead, this method creates a copy of est internally, where est.base is replaced by measure.e.base. Therefore, use measure to control the "unit" of the mutual information.

Estimators

Estimator	Principle	CMIShannon

Kraskov	Nearest neighbors	✓

Zhu	Nearest neighbors	✓

Gao	Nearest neighbors	✓

Goria	Nearest neighbors	✓

Lord	Nearest neighbors	✓

LeonenkoProzantoSavani	Nearest neighbors	✓

condmutualinfo([measure::CMI], est::MutualInformationEstimator, x, y, z) → cmi::Real

Estimate the mutual information between x and y conditioned on z, using the given conditional mutual information (CMI) measure, computed as a a difference of mutual information terms (just the chain rule of mutual information)

\[\hat{I}(X; Y | Z) = \hat{I}(X; Y, Z) - \hat{I}(X; Z).\]

The MutualInformationEstimator est may be continuous/differential, discrete or mixed. No bias correction in performed, except the bias correction that occurs for each individual mutual information term. If measure is not given, then the default is CMIShannon().

Estimators

Estimator	Type	Principle	CMIShannon

KraskovStögbauerGrassberger1	Continuous	Nearest neighbors	✓

KraskovStögbauerGrassberger2	Continuous	Nearest neighbors	✓

GaoKannanOhViswanath	Mixed	Nearest neighbors	✓

GaoOhViswanath	Continuous	Nearest neighbors	✓

Definitions

CausalityTools.ConditionalMutualInformation — Type

ConditionalMutualInformation <: AssociationMeasure
CMI # alias

The supertype of all conditional mutual information measures. Concrete subtypes are

CMIShannon

CMIRenyiJizba

CMIRenyiPoczos

ConditionalMutualInformationEstimators

CausalityTools.ConditionalMutualInformationEstimator — Type

ConditionalMutualInformationEstimator <: InformationEstimator
CMIEstimator # alias

The supertype of all conditional mutual information estimators.

Subtypes

FPVP.

PoczosSchneiderCMI.

Rahimzamani.

MesnerShalisi.

GaussianCMI (parametric)

CausalityTools.GaussianCMI — Type

GaussianCMI <: MutualInformationEstimator
GaussianCMI(; normalize::Bool = false)

GaussianCMI is a parametric estimator for Shannon conditional mutual information (CMI) (Vejmelka & Paluš)[Vejmelka2008].

Description

GaussianCMI estimates Shannon CMI through a sum of two mutual information terms that each are estimated using GaussianMI (the normalize keyword is the same as for GaussianMI):

\[\hat{I}_{Gaussian}(X; Y | Z) = \hat{I}_{Gaussian}(X; Y, Z) - \hat{I}_{Gaussian}(X; Z)\]

FPVP

CausalityTools.FPVP — Type

FPVP <: ConditionalMutualInformationEstimator
FPVP(k = 1, w = 0)

The Frenzel-Pompe-Vejmelka-Paluš (or FPVP for short) estimator is used to estimate the differential conditional mutual information using a k-th nearest neighbor approach that is analogous to that of the KraskovStögbauerGrassberger1 mutual information estimator (Frenzel & Pompe, 2007[Frenzel2007]; Vejmelka & Paluš, 2008[Vejmelka2008]).

w is the Theiler window, which controls the number of temporal neighbors that are excluded during neighbor searches.

MesnerShalisi

CausalityTools.MesnerShalisi — Type

MesnerShalisi <: ConditionalMutualInformationEstimator
MesnerShalisi(k = 1, w = 0)

The MesnerShalisi estimator is an estimator for conditional mutual information for data that can be mixtures of discrete and continuous data (Mesner & Shalisi et al., 2020)[MesnerShalisi2020].

PoczosSchneiderCMI

CausalityTools.PoczosSchneiderCMI — Type

PoczosSchneiderCMI <: ConditionalMutualInformationEstimator
PoczosSchneiderCMI(k = 1, w = 0)

The PoczosSchneiderCMI estimator computes various (differential) conditional mutual informations, using a k-th nearest neighbor approach (Póczos & Schneider, 2012)[Póczos2012].

Rahimzamani

CausalityTools.Rahimzamani — Type

Rahimzamani <: ConditionalMutualInformationEstimator
Rahimzamani(k = 1, w = 0)

The Rahimzamani estimator, short for Rahimzamani-Asnani-Viswanath-Kannan, is an estimator for Shannon conditional mutual information for data that can be mixtures of discrete and continuous data (Rahimzamani et al., 2018)[Rahimzamani2018].

This is very similar to the GaoKannanOhViswanath mutual information estimator, but has been expanded to the conditional case.

Vejmelka2008Vejmelka, M., & Paluš, M. (2008). Inferring the directionality of coupling with conditional mutual information. Physical Review E, 77(2), 026214.

Frenzel2007Frenzel, S., & Pompe, B. (2007). Partial mutual information for coupling analysis of multivariate time series. Physical review letters, 99(20), 204101. w is the Theiler window.

Vejmelka2008Vejmelka, M., & Paluš, M. (2008). Inferring the directionality of coupling with conditional mutual information. Physical Review E, 77(2), 026214.

MesnerShalisi2020Mesner, O. C., & Shalizi, C. R. (2020). Conditional mutual information estimation for mixed, discrete and continuous data. IEEE Transactions on Information Theory, 67(1), 464-484.

Póczos2012Póczos, B., & Schneider, J. (2012, March). Nonparametric estimation of conditional information and divergences. In Artificial Intelligence and Statistics (pp. 914-923). PMLR.

Rahimzamani2018Rahimzamani, A., Asnani, H., Viswanath, P., & Kannan, S. (2018). Estimators for multivariate information measures in general probability spaces. Advances in Neural Information Processing Systems, 31.