Part mutual information API

The part mutual information (PMI) API is defined by

PMI, and its subtypes.

pmi,

CausalityTools.pmi — Function

pmi([measure::CMI], est::ProbabilitiesEstimator, x, y, z) → pmi_est::Real ∈ [0, a)

Estimate the part mutual information (PMI; Zhao et al. (2016)[Zhao2016]).

If measure is not given, then the default is PMI(; base = 2). With a ProbabilitiesEstimator, the returned pmi_est is guaranteed to be non-negative.

Estimators

Estimator	Principle	PMI

CountOccurrences	Frequencies	✓

ValueHistogram	Binning (histogram)	✓

SymbolicPermutation	Ordinal patterns	✓

Dispersion	Dispersion patterns	✓

Zhao2016Zhao, J., Zhou, Y., Zhang, X., & Chen, L. (2016). Part mutual information for quantifying direct associations in networks. Proceedings of the National Academy of Sciences, 113(18), 5130-5135.