Recurrence API

CausalityTools.mcr — Function

mcr(m::MCR, x, y)

Compute the association between x and y based on conditional probabilities of recurrence using the given MCR measure, where x and y can be either univariate timeseries or multivariate StateSpaceSets.

CausalityTools.rmcd — Function

rmcd(measure::RMCD, x, y)
rmcd(measure::RMCD, x, y, [z, ...])

Estimate the recurrence-based measure of dependence between x and y, conditional on z if given.

Parameters for recurrence matrix estimation are given as a RMCD instance. Inputs x, y, z can be either univariate timeseries or multivariate StateSpaceSets.