Interventional complexity causality (ICC)

Interface

CausalityTools.InterventionalComplexityCausality.icc — Function

icc(x::Vector{Int}, y::Vector{Int}, estimator)

Interventional complexity causality (ICC) is defined as[Kathpalia2019] the change in the dynamical complexity of time series X when Xⁱ is seen to be generated jointly by the dynamical evolution of both Y⁻ and X⁻ as opposed to by the reality of the dynamical evolution of X⁻ alone", that is

\[ICC_{Y^{-} \to X^{i}} = DC(X^{i} | X^{i}) - DC(X^{i} | X^{-}, Y^{-}),\]

where

DC is dynamical_complexity.

Xⁱ is a block of current values of X (ΔX in the original paper),

X⁻ is a block of past values of X, not overlapping with and immediately before Xⁱ, and

Y⁻ is a block of values from a time series Y taken at the same indices as X⁻.

icc computes the mean ICC across a series of sliding windows over the input time series x and y, where the sliding window configuration is dictated by the estimator.

Interpretation

If $ICC_{Y \to X}$ is statistically zero, then it implies no causal influence from $Y$ to $X$. If $ICC_{Y \to X}$ is statistically different from zero, then it is inferred that $Y$ dynamically influences $X$, and a higher magnitudes indicates higher degrees of causation.

$ICC_{Y \to X}$ can be both positive and negative. For details on interpretating both signs, see Kathpalia and Nagaraj (2019)[Kathpalia2019].

Examples

using Statistics, Random
rng = MersenneTwister(1234)
x = rand(rng, 0:1, 500)
y = rand(rng, 0:1, 500)
est = CompressionComplexityCausalityEstimator(
    algorithm = EffortToCompress(normalize = true),
    w = 15,
    L = 30,
    step = 10)
icc(x, y, est)

# output
0.10375494071146242

See also: CompressionComplexityCausalityEstimator, EffortToCompress

Estimators

CausalityTools.InterventionalComplexityCausality.CompressionComplexityCausalityEstimator — Type

CompressionComplexityCausalityEstimator(
    algorithm::CompressionComplexityAlgorithm = EffortToCompress(),
    w::In = 15,
    L::Int = 30)

CompressionComplexityCausalityEstimator (CCCEstimator for short) stores parameters used to calculate compute interventional complexity causality (ICC):.

Parameters

algorithm: The algorithm used to compute the compression complexity, e.g. an instance of EffortToCompress.

w: The block size for Xⁱ, which is a block of current values of time series X. The default value is set to 15, as recommended in Kathpalia & Nagaraj (2019).

L: The block size for X⁻, Y⁻, Z⁻, ... , which are blocks of past values of X, Y, Z, and so on, whose values are taken from the immediate past of the Xⁱ block. The default value is L = 30, but for real applications, this value should determined as described in table S2 in the supplement of Kathpalia & Nagaraj (2019).

step: ICC is computed over a sliding window of width w + L. step gives the shift, in number of indices, between windows ($\delta$ in Kathpalia & Nagaraj, 2019).

Input data requirements

All estimators assume that the input time series are pre-symbolized/binned integer-valued time series. Thus, the parameter B (number of bins) from Kathpalia & Nararaj (2019)[Kathpalia2019] is not present here.

See also: icc, EffortToCompress.

Examples

Kathpalia2019Kathpalia, A., & Nagaraj, N. (2019). Data-based intervention approach for Complexity-Causality measure. PeerJ Computer Science, 5, e196.

Kathpalia2019Kathpalia, A., & Nagaraj, N. (2019). Data-based intervention approach for Complexity-Causality measure. PeerJ Computer Science, 5, e196.