Conditional mutual information (CMI)

API

The condition mutual information API is defined by

ConditionalMutualInformation,

mutualinfo,

ConditionalMutualInformationEstimator.

Definitions

Shannon CMI

CausalityTools.ConditionalMutualInformation — Type

ConditionalMutualInformation <: InformationMeasure
CMI # alias

The supertype of all conditional mutual informations.

CausalityTools.CMIShannon — Type

CMIShannon <: ConditionalMutualInformation
CMIShannon(; base = 2)

The Shannon conditional mutual information (CMI) $I^S(X; Y | Z)$.

Supported definitions

Consider random variables $X \in \mathbb{R}^{d_X}$ and $Y \in \mathbb{R}^{d_Y}$, given $Z \in \mathbb{R}^{d_Z}$. The Shannon conditional mutual information is defined as

\[\begin{align*} I(X; Y | Z) &= H^S(X, Z) + H^S(Y, z) - H^S(X, Y, Z) - H^S(Z) \\ &= I^S(X; Y, Z) + I^S(X; Y) \end{align*},\]

where $I^S(\cdot; \cdot)$ is the Shannon mutual information MIShannon, and $H^S(\cdot)$ is the Shannon entropy.

Differential Shannon CMI is obtained by replacing the entropies by differential entropies.

See also: condmutualinfo.

CausalityTools.CMIRenyiJizba — Type

CMIRenyiJizba <: ConditionalMutualInformation

The Rényi conditional mutual information $I_q^{R_{J}}(X; Y | Z$ defined in Jizba et al. (2012)[Jizba2012].

Definition

\[I_q^{R_{J}}(X; Y | Z) = I_q^{R_{J}}(X; Y, Z) - I_q^{R_{J}}(X; Z),\]

where $I_q^{R_{J}}(X; Z)$ is the MIRenyiJizba mutual information.

CausalityTools.CMIRenyiPoczos — Type

CMIRenyiPoczos <: ConditionalMutualInformation

The differential Rényi conditional mutual information $I_q^{R_{P}}(X; Y | Z)$ defined in (Póczos & Schneider, 2012)[Póczos2012].

Definition

\[\begin{align*} I_q^{R_{P}}(X; Y | Z) = \dfrac{1}{q-1} \int \int \int \dfrac{p_Z(z) p_{X, Y | Z}^q}{( p_{X|Z}(x|z) p_{Y|Z}(y|z) )^{q-1}} \\ \mathbb{E}_{X, Y, Z} \sim p_{X, Y, Z} \left[ \dfrac{p_{X, Z}^{1-q}(X, Z) p_{Y, Z}^{1-q}(Y, Z) }{p_{X, Y, Z}^{1-q}(X, Y, Z) p_Z^{1-q}(Z)} \right] \end{align*}\]

Dedicated estimators

CausalityTools.condmutualinfo — Method

condmutualinfo([measure::CMI], est::CMIEstimator, x, y, z) → cmi::Real

Estimate a conditional mutual information (CMI) of some kind (specified by measure), between x and y, given z, using the given dedicated ConditionalMutualInformationEstimator, which may be discrete, continuous or mixed.

CausalityTools.ConditionalMutualInformationEstimator — Type

ConditionalMutualInformationEstimator <: InformationEstimator
CMIEstimator # alias

The supertype of all conditional mutual information estimators.

Subtypes

FPVP.

PoczosSchneiderCMI.

Rahimzamani.

MesnerShalisi.

CausalityTools.FPVP — Type

FPVP <: ConditionalMutualInformationEstimator
FPVP(k = 1, w = 0)

The Frenzel-Pompe-Vejmelka-Paluš (or FPVP for short) estimator is used to estimate the differential conditional mutual information using a k-th nearest neighbor approach that is analogous to that of the KraskovStögbauerGrassberger1 mutual information estimator (Frenzel & Pompe, 2007[Frenzel2007]; Vejmelka & Paluš, 2008[Vejmelka2008]).

w is the Theiler window, which controls the number of temporal neighbors that are excluded during neighbor searches.

CausalityTools.MesnerShalisi — Type

MesnerShalisi <: ConditionalMutualInformationEstimator
MesnerShalisi(k = 1, w = 0)

The MesnerShalisi estimator is an estimator for conditional mutual information for data that can be mixtures of discrete and continuous data (Mesner & Shalisi et al., 2020)[MesnerShalisi2020].

CausalityTools.PoczosSchneiderCMI — Type

PoczosSchneiderCMI <: ConditionalMutualInformationEstimator
PoczosSchneiderCMI(k = 1, w = 0)

The PoczosSchneiderCMI estimator computes various (differential) conditional mutual informations, using a k-th nearest neighbor approach (Póczos & Schneider, 2012)[Póczos2012].

CausalityTools.Rahimzamani — Type

Rahimzamani <: ConditionalMutualInformationEstimator
Rahimzamani(k = 1, w = 0)

The Rahimzamani estimator, short for Rahimzamani-Asnani-Viswanath-Kannan, is an estimator for Shannon conditional mutual information for data that can be mixtures of discrete and continuous data (Rahimzamani et al., 2018)[Rahimzamani2018].

This is very similar to the GaoKannanOhViswanath mutual information estimator, but has been expanded to the conditional case.

Table of dedicated CMI estimators

Estimator	Principle	CMIShannon	CMIRenyiPoczos

FPVP	Nearest neighbors	✓	x

MesnerShalisi	Nearest neighbors	✓	x

Rahimzamani	Nearest neighbors	✓	x

PoczosSchneiderCMI	Nearest neighbors	x	✓

GaussianCMI	Parametric	✓	x

Estimation through mutual information

CausalityTools.condmutualinfo — Method

condmutualinfo([measure::CMI], est::MutualInformationEstimator, x, y, z) → cmi::Real

Estimate the conditional mutual information (CMI) measure between x and y using a difference of mutual information terms, without any bias correction, using the provided MutualInformationEstimator est, which may be continuous/differential, discrete or mixed. If measure is not given, then the default is CMIShannon().

Estimator	Type	Principle	CMIShannon

KraskovStögbauerGrassberger1	Continuous	Nearest neighbors	✓

KraskovStögbauerGrassberger2	Continuous	Nearest neighbors	✓

GaoKannanOhViswanath	Mixed	Nearest neighbors	✓

GaoOhViswanath	Continuous	Nearest neighbors	✓

GaussianMI		Parametric	✓

Discrete CMI

CausalityTools.condmutualinfo — Method

condmutualinfo([measure::CMI], est::ProbabilitiesEstimator, x, y, z) → cmi::Real ∈ [0, a)

Estimate the conditional mutual information (CMI) measure between x and y given z using a sum of entropy terms, without any bias correction, using the provided ProbabilitiesEstimator est. If measure is not given, then the default is CMIShannon().

With a ProbabilitiesEstimator, the returned cmi is guaranteed to be non-negative.

Table of discrete mutual information estimators

Here, we list the ProbabilitiesEstimators that are compatible with condmutualinfo, and which definitions they are valid for.

Estimator	Principle	CMIShannon	CMIRenyiSarbu

CountOccurrences	Frequencies	✓	✓

ValueHistogram	Binning (histogram)	✓	✓

SymbolicPermuation	Ordinal patterns	✓	✓

Dispersion	Dispersion patterns	✓	✓

Differential CMI

CausalityTools.condmutualinfo — Method

condmutualinfo([measure::CMI], est::DifferentialEntropyEstimator, x, y, z) → cmi

Estimate the conditional mutual information (CMI) measure between x and y using a sum of entropy terms, without any bias correction, using the provided DifferentialEntropyEstimator est (which must support multivariate data). If measure is not given, then the default is CMIShannon().

Estimator	Principle	Input data	CMIShannon

Kraskov	Nearest neighbors	Dataset	✓

Zhu	Nearest neighbors	Dataset	✓

Gao	Nearest neighbors	Dataset	✓

Goria	Nearest neighbors	Dataset	✓

Lord	Nearest neighbors	Dataset	✓

LeonenkoProzantoSavani	Nearest neighbors	Dataset	✓

Jizba2012Jizba, P., Kleinert, H., & Shefaat, M. (2012). Rényi’s information transfer between financial time series. Physica A: Statistical Mechanics and its Applications, 391(10), 2971-2989.

Póczos2012Póczos, B., & Schneider, J. (2012, March). Nonparametric estimation of conditional information and divergences. In Artificial Intelligence and Statistics (pp. 914-923). PMLR.

Frenzel2007Frenzel, S., & Pompe, B. (2007). Partial mutual information for coupling analysis of multivariate time series. Physical review letters, 99(20), 204101. w is the Theiler window.

Vejmelka2008Vejmelka, M., & Paluš, M. (2008). Inferring the directionality of coupling with conditional mutual information. Physical Review E, 77(2), 026214.

MesnerShalisi2020Mesner, O. C., & Shalizi, C. R. (2020). Conditional mutual information estimation for mixed, discrete and continuous data. IEEE Transactions on Information Theory, 67(1), 464-484.

Póczos2012Póczos, B., & Schneider, J. (2012, March). Nonparametric estimation of conditional information and divergences. In Artificial Intelligence and Statistics (pp. 914-923). PMLR.

Rahimzamani2018Rahimzamani, A., Asnani, H., Viswanath, P., & Kannan, S. (2018). Estimators for multivariate information measures in general probability spaces. Advances in Neural Information Processing Systems, 31.