Correlation measures

Pearson correlation

CausalityTools.PearsonCorrelation — Type

PearsonCorrelation

The Pearson correlation of two variables.

Usage

Use with independence to perform a formal hypothesis test for pairwise dependence.

Use with pearson_correlation to compute the raw correlation coefficient.

Description

The sample Pearson correlation coefficient for real-valued random variables $X$ and $Y$ with associated samples $\{x_i\}_{i=1}^N$ and $\{y_i\}_{i=1}^N$ is defined as

\[\rho_{xy} = \dfrac{\sum_{i=1}^n (x_i - \bar{x})(y_i - \bar{y}) }{\sqrt{\sum_{i=1}^N (x_i - \bar{x})^2}\sqrt{\sum_{i=1}^N (y_i - \bar{y})^2}},\]

where $\bar{x}$ and $\bar{y}$ are the means of the observations $x_k$ and $y_k$, respectively.

CausalityTools.pearson_correlation — Function

pearson_correlation(x::VectorOrDataset, y::VectorOrDataset)

Compute the PearsonCorrelation between x and y, which must each be 1-dimensional.

Partial correlation

CausalityTools.PartialCorrelation — Type

PartialCorrelation

The correlation of two variables, with the effect of a set of conditioning variables removed.

Usage

Use with independence to perform a formal hypothesis test for conditional dependence.

Use with partial_correlation to compute the raw correlation coefficient.

Description

There are several ways of estimating the partial correlation. We follow the matrix inversion method, because for Datasets, we can very efficiently compute the required joint covariance matrix $\Sigma$ for the random variables.

Formally, let $X_1, X_2, \ldots, X_n$ be a set of $n$ real-valued random variables. Consider the joint precision matrix,$\P = (p_{ij}) = \Sigma^-1$. The partial correlation of any pair of variables $(X_i, X_j)$, given the remaining variables $\bf{Z} = \{X_k\}_{i=1, i \neq i, j}^n$, is defined as

\[\rho_{X_i X_j | \bf{Z}} = -\dfrac{p_ij}{\sqrt{ p_{ii} p_{jj} }}\]

In practice, we compute the estimate

\[\hat{\rho}_{X_i X_j | \bf{Z}} = -\dfrac{\hat{p}_ij}{\sqrt{ \hat{p}_{ii} \hat{p}_{jj} }},\]

where $\hat{P} = \hat{\Sigma}^{-1}$ is the sample precision matrix.

CausalityTools.partial_correlation — Function

partial_correlation(x::VectorOrDataset, y::VectorOrDataset,
    z::VectorOrDataset...)

Compute the PartialCorrelation between x and y, given z.

Distance correlation

CausalityTools.DistanceCorrelation — Type

DistanceCorrelation

The distance correlation (Székely et al., 2007)[Székely2007] measure quantifies potentially nonlinear associations between pairs of variables.

Usage

Use with independence to perform a formal hypothesis test for pairwise dependence.

Use with distance_correlation to compute the raw distance correlation coefficient.

CausalityTools.distance_correlation — Function

distance_correlation(x, y) → dcor ∈ [0, 1]

Compute the empirical/sample distance correlation (Székely et al., 2007)[Székely2007], here called dcor, between datasets x and y.

Székely2007Székely, G. J., Rizzo, M. L., & Bakirov, N. K. (2007). Measuring and testing dependence by correlation of distances. The annals of statistics, 35(6), 2769-2794.