Experimental

Here we list implemented methods that do not yet appear in peer-reviewed journals, but are found, for example, in pre-print servers like arXiv.

The API for these methods, and their return values, may change at any time until they appear as part of the public API. Use them wisely.

Predictive asymmetry

CausalityTools.predictive_asymmetry — Function

predictive_asymmetry(estimator::TransferEntropyEstimator, ηs; s, t, [c],
    dTf = 1, dT = 1, dS = 1, τT = -1, τS = -1, [dC = 1, τC = -1],
    normalize::Bool = false, f::Real = 1.0, base = 2) → Vector{Float64}

Compute the predictive asymmetry[Haaga2020] 𝔸(s → t) for source time series s and target time series t over prediction lags ηs, using the given estimator and embedding parameters dTf, dT, dS, τT, τS (see also EmbeddingTE)

If a conditional time series c is provided, compute 𝔸(s → t | c). Then, dC and τC controls the embedding dimension and embedding lag for the conditional variable.

Returns

Returns a vector containing the predictive asymmetry for each value of ηs.

Normalization (hypothesis test)

If normalize == true (the default), then compute the normalized predictive asymmetry 𝒜. In this case, for each $\eta$ in ηs, compute 𝒜(η) by normalizing 𝔸(η) to some fraction f of the mean transfer entropy over prediction lags $-\eta, ..., \eta$ (exluding lag 0). Haaga et al. (2020)[Haaga2020] uses a normalization with f=1.0 as a built-in hypothesis test, avoiding more computationally costly surrogate testing.

Estimators

Any estimator that works for transferentropy will also work with predictive_asymmetry. Check the online documentation for compatiable estimators.

Examples

using CausalityTools
# Some example time series
x, y = rand(100), rand(100)
# 𝔸(x → y) over prediction lags 1:5
𝔸reg  = predictive_asymmetry(x, y, VisitationFrequency(RectangularBinning(5)), 1:5)

Experimental!

This is a method that does not yet appear in a peer-reviewed scientific journal. Feel free to use, but consider it experimental for now. It will reappear in a 2.X release in new form once published in a peer-reviewed journal.

Independence testing

A common application of information theoretic methods such as conditional mutual information (condmutualinfo) is in the context of null hypothesis testing for the conditional independence of variables.

Depending on the context, the input data and the method used, there are many considerations to be made about how to perform this conditional independence testing. Luckily, many excellent frameworks for doing so exist in the literature.

Here, we present some commonly used independence tests from the scientific literature, which can all be seamlessly used with the function independence, with any measure that quantifies conditional independence, in combination with any compatible estimator.

For example, in just a few lines of code, you can perform Runge's local permutation (LocalPermutation test on your data with over 20 different estimators for the conditional mutual information. If your application rather calls for the use of traditional surrogate data, the SurrogateCIT test seamlessly integrates with any time series surrogate method from TimeseriesSurrogates.jl.

Independence test API

The independence test API is defined by

independence

IndependenceTest

ConditionalIndependenceTest

Independence tests

CausalityTools.independence — Function

independence(test, x, y, [z])

Perform the given test of independence between x and y using the provided test. If z is given, compute the conditional independence of x and y given z.

This function is just a generic implementation of a one-sided hypothesis test, where the null hypothesis is that x and y are independent (given z, if provided).

Supported tests

The null hypothesis is specified by test, which is a IndependenceTest.

LocalPermutation

SurrogateCIT. This is essentially a convenience wrapper that performs standard surrogate testing.

CausalityTools.ConditionalIndependenceTest — Type

ConditionalIndependenceTest <: IndependenceTest

The supertype for all conditional independence tests, which are:

LocalPermutation.

Surrogate test (global permutation)

CausalityTools.SurrogateCIT — Type

SurrogateCIT <: IndependenceTest
SurrogateCIT(;
    measure = CMIShannon(),
    est = FPVP(k = 5),
    nsurr::Int = 100,
    surrogate = RandomShuffle(),
    rng = Random.MersenneTwister(1234),
)

The SurrogateCIT test is a generic conditional independence test (CIT) for assessing whether two variables X and Y are conditionally independendent given a third variable Z. It uses surrogate time series to generate the null distribution.

Null hypotheses

If used with a ConditionalMutualInformation measure such as CMIShannon, then the shuffled variable is the first input X.

If used with a TransferEntropy measure such as TEShannon, then the shuffled variable is the first input X, i.e. the source variable.

Example usage

Here, we'll test if the transfer entropy from x to y is significant. If we can reject the null hypothesis that x and y are independent, then we take that as evidence that x influences y.

using CausalityTools
# If `x` and `y` were dependent given `z`, then we'd expect to be able to reject the
# null hypothesis that x ⫫ y | z.
sys = logistic2_unidir(c_xy = 0.5)
npts = 1000
x, y = columns(trajectory(sys, npts, Ttr = npts*10))
test = SurrogateCIT(TEShannon(), FPVP(k = 10))
independence(test, x, y)

# If `x` and `y` are *independent*.

Local permutation

CausalityTools.LocalPermutation — Type

LocalPermutation <: ConditionalIndependenceTest
LocalPermutation(measure = CMIShannon(), est = FPVP(k = 5);
    kperm::Int = 5,
    nsurr::Int = 100,
    rng = Random.MersenneTwister(1234))

A generic implementation of the LocalPermutation test (Runge, 2018)[Runge2018], which tests whether two variables X and Y are conditionally independendent given a third variable Z (all of which may be multivariate).

You can use any valid combination of definition, measure and est that yields a conditional dependence measure $\hat{M}(X; Y | Z)$ with the same ordering as condmutualinfo (i.e. the conditional variable (Z) is the third input variable). The default measure is Shannon conditional mutual information CMIShannon, $I(X; Y | Z)$ with the nearest-neighbor based FPVP differential CMI estimator.

Description

LocalPermutation creates permuted X values using a local permutation scheme that is based on kperm-th nearest neighbor searches. Permuted points are constructed as $(x_i^*, y_i, z_i)_{i=1}^N$, where the goal is that $x_i^*$ are drawn without replacement, and $x_i$ is replaced by $x_j$ only if $z_i \approx z_j$. Then, for each permuted version of X, it computes the original statistic on the permuted data, keeping Y and Z fixed, i.e. $\hat{M}(\hat{X}; Y | Z)$.

Example usage

x = randn(2000)
y = randn(2000) .+ 0.7 .* x
z = sin.(randn(2000)) .* 0.5 .* y
test = LocalPermutation(CMIShannon(; base = 2), FPVP(k = 10))
independence(test, x, y, z)

Examples

LocalPermutation

Here, we'll create a three-variable scenario where X and Z are connected through Y, so that $I(X; Z | Y) = 0$ and $I(X; Y | Z) > 0$. We'll test for conditional independence using Shannon conditional mutual information (CMIShannon). To estimate CMI, we'll use the Kraskov differential entropy estimator, which naively computes CMI as a sum of entropy terms without guaranteed bias cancellation.

using CausalityTools

X = randn(1000)
Y = X .+ randn(1000) .* 0.4
Z = randn(1000) .+ Y
x, y, z = Dataset.((X, Y, Z))
test = LocalPermutation(CMIShannon(base = 2), Kraskov(k = 10), nsurr = 30)
test_result = independence(test, x, y, z)

`LocalPermutation` independence test
----------------------------------------------------------------------------------
H₀: "The first two variables are conditionally independent given the third"
----------------------------------------------------------------------------------
Estimated: 0.432551591633928
Ensemble quantiles (30 permutations):
  (99.9%): 0.004668358065417452
  (99%):   0.004401725585744441
  (95%):   0.0011608347945485636
p-value:   0.0
  H₀ rejected at α = 0.05:  Yes ✅
  H₀ rejected at α = 0.01:  Yes ✅
  H₀ rejected at α = 0.001: Yes ✅

We expect there to be a detectable influence from $X$ to $Y$, if we condition on $Z$ or not, because $Z$ doesn't influence neither $X$ nor $Y$. The null hypothesis is that the first two variables are conditionally independent given the third, which we reject with a very low p-value. Hence, we accept the alternative hypothesis that the first two variables $X$ and $Y$. are conditionally dependent given $Z$.

test_result = independence(test, x, z, y)

`LocalPermutation` independence test
----------------------------------------------------------------------------------
H₀: "The first two variables are conditionally independent given the third"
----------------------------------------------------------------------------------
Estimated: -0.024075254737095264
Ensemble quantiles (30 permutations):
  (99.9%): 0.0016749342584386241
  (99%):   0.00023808299282294335
  (95%):   -0.004618503528311589
p-value:   0.6666666666666666
  H₀ rejected at α = 0.05:  No  ❌
  H₀ rejected at α = 0.01:  No  ❌
  H₀ rejected at α = 0.001: No  ❌

As expected, we cannot reject the null hypothesis that $X$ and $Z$ are conditionally independent given $Y$, because $Y$ is the variable that transmits information from $X$ to $Z$.

SurrogateCIT

To demonstrate the SurrogateCIT test, we use the transfer entropy measure, which accepts either two input timeseries, or three timeseries when computing the partial/conditional transfer entropy.

using CausalityTools
sys = logistic2_unidir(c_xy = 0.5) # x affects y, but not the other way around.
x, y = columns(trajectory(sys, 1000, Ttr = 10000))

test = SurrogateCIT(TEShannon(), KSG1(k = 4))
independence(test, x, y)

`SurrogateCIT` independence test result
-----------------------------------------------------------------------------------
H₀: "The first two variables are independent (given the 3rd variable, if relevant)"
Hₐ: "The first two variables are dependent (given the 3rd variable, if relevant)"
-----------------------------------------------------------------------------------
Estimated: 1.2878586270076275
Ensemble quantiles (100 permutations):
  (99.9%): 0.6890197450381296
  (99%):   0.6786060026795824
  (95%):   0.6673634762250779
p-value:   0.0
  H₀ rejected at α = 0.05:  Yes ✅
  H₀ rejected at α = 0.01:  Yes ✅
  H₀ rejected at α = 0.001: Yes ✅

As expected, we can reject the null hypothesis that the future of y is independent of x, because x does actually influence y. This doesn't change if we compute partial transfer entropy with respect to some random extra time series, because it doesn't influence either variables.

independence(test, x, y, rand(length(x)))

`SurrogateCIT` independence test result
-----------------------------------------------------------------------------------
H₀: "The first two variables are independent (given the 3rd variable, if relevant)"
Hₐ: "The first two variables are dependent (given the 3rd variable, if relevant)"
-----------------------------------------------------------------------------------
Estimated: 0.5593817615003804
Ensemble quantiles (100 permutations):
  (99.9%): 0.5112567506764502
  (99%):   0.506537007792058
  (95%):   0.49438932005156394
p-value:   0.0
  H₀ rejected at α = 0.05:  Yes ✅
  H₀ rejected at α = 0.01:  Yes ✅
  H₀ rejected at α = 0.001: Yes ✅

Automated embedding for transfer entropy

Missing docstring.

Missing docstring for bbnue. Check Documenter's build log for details.

Haaga2020Haaga, Kristian Agasøster, David Diego, Jo Brendryen, and Bjarte Hannisdal. "A simple test for causality in complex systems." arXiv preprint arXiv:2005.01860 (2020).

Runge2018Runge, J. (2018, March). Conditional independence testing based on a nearest-neighbor estimator of conditional mutual information. In International Conference on Artificial Intelligence and Statistics (pp. 938-947). PMLR.