Joint distance distribution

CausalityTools.JointDistanceDistribution — Type

JointDistanceDistribution <: AssociationMeasure end
JointDistanceDistribution(; metric = Euclidean(), B = 10, D = 2, τ = 1, μ = 0.0)

The joint distance distribution (JDD) measure (Amigó & Hirata, 2018)[Amigo2018].

Keyword arguments

distance_metric::Metric: An instance of a valid distance metric from Distances.jl. Defaults to Euclidean().

B::Int: The number of equidistant subintervals to divide the interval [0, 1] into when comparing the normalised distances.

D::Int: Embedding dimension.

τ::Int: Embedding delay.

μ: The hypothetical mean value of the joint distance distribution if there is no coupling between x and y (default is μ = 0.0).

Independence testing

If used with independence, then a HypothesisTests.OneSampleTTest is returned. Use pvalue with keyword tail = :right on the result to get the p-value at 95% confidence.

Examples

Computing the JJD

Independence testing using JJD

CausalityTools.jdd — Function

jdd(measure::JointDistanceDistribution, source, target) → Δ

Compute the joint distance distribution (Amigó & Hirata, 2018[Amigo2018]) from source to target using the given JointDistanceDistribution measure.

Returns the distribution Δ from the paper directly (example). Use JointDistanceDistributionTest to perform a formal indepencence test.

Amigo2018Amigó, José M., and Yoshito Hirata. "Detecting directional couplings from multivariate flows by the joint distance distribution." Chaos: An Interdisciplinary Journal of Nonlinear Science 28.7 (2018): 075302.

Amigo2018Amigó, José M., and Yoshito Hirata. "Detecting directional couplings from multivariate flows by the joint distance distribution." Chaos: An Interdisciplinary Journal of Nonlinear Science 28.7 (2018): 075302.