Penchants and leanings

Penchants and leanings are probability-based exploratory causal inference measures that rely on the estimation of probabilities directly from the data. Here, we provide implementations of the penchants and leanings algoriths from McCracken and Weigel (2016).

CausalityTools.Leanings.penchant — Function

penchant(x, y, l; weighted = false) → ρ̄

Computes the mean observed penchant ρ̄ (McCracken & Weigel, 2016)[McCrackenWeigel2016] of the l-assignment $\{\text{C}, \text{E}\} = \{\text{cause}, \text{effect}\} =\{x_{t-l}, y_t\}$.

If weighted == true, then compute the weighted mean observed penchant.

Definition

The causal penchant , or causal tendency, is a causal indicator defined as $\rho_{CE} = P(E|C) - P(E|\bar{C})$, where $\rho_{CE} \in [-1, 1]$ (McCracken & Weigel, 2016). If $\rho_{CE} > 0$, then $C$ causes or drives $E$, and if $\rho_{CE} \leq 0$, then the direction of influence is undetermined.

A direct formula for $\rho_{CE}$ can be obtained using Bayes' theorem.

\[P(E|C) = P(C|E) \dfrac{P(E)}{P(C)} \]

Using the definitions of probability complements, one arrives at the following expression (see the original paper for a detailed derivation):

\[\rho_{CE} = P(E|C) \left[1 + \dfrac{P(C)}{1-P(C)} \right] - \dfrac{P(E)}{1 - P(C)}.\]

Applying appropriate discretization schemes, these probabilities can be estimated directly from time series using simple counting, which makes the method fast and well-suited for exploratory causal inference analysis.

See also lean and data requirements discussed therein.

CausalityTools.Leanings.lean — Function

lean(x, y, l = 1; weighted = true) → ρ̄

Compute the mean observed leaning λ ∈ [-2, 2] (McCracken & Weigel, 2016)[McCrackenWeigel2016] between scalar time series x and y.

If ρ̄ > 0, then the probability that x drives y is higher than the probability that y drives x. Vice versa, if ρ̄ > 0, then the probability that y drives x is higher than the probability that x drives y.

Weighting

If weighted == true, then each penchant is weighted by the number of times it appears in the data, which yields the weighted mean observed leaning. This the default behavior, since "...[the weighted mean observed leaning] accounts for the frequency of observed cause-effect pairs within the data, which is assumed to be a predictor of causal influence" (McCracken & Weigel, 2016).

Data requirements

Both x and y must be discrete, and the number discrete states can't be too high relative to the number of points in x and y. Otherwise, probabilities cannot be reliably estimated. If the function returns NaN, then you probably either haven't discretized your time series, or the partition is too fine-grained given the numbers of points available.

Hence, this implementation does not take care of the tolerance domains discussed in the paper. Pre-process your time series using appropriate binning or symbolization schemes.

McCrackenWeigel2016McCracken, J. M., & Weigel, R. S. (2016). Nonparametric causal inference for bivariate time series. Physical Review E, 93(2), 022207.

McCrackenWeigel2016McCracken, J. M., & Weigel, R. S. (2016). Nonparametric causal inference for bivariate time series. Physical Review E, 93(2), 022207.