Inferring causal graphs

Causal graph API

The API for inferring causal graphs is defined by:

infer_graph

GraphAlgorithm, and its subtypes

CausalityTools.infer_graph — Function

infer_graph(algorithm::GraphAlgorithm, x) → g

Infer graph from input data x using the given algorithm.

Returns g, whose type depends on algorithm.

CausalityTools.GraphAlgorithm — Type

GraphAlgorithm

The supertype of all causal graph inference algorithms.

Concrete implementations

OCE. The optimal causation entropy algorithm for time series graphs.

Optimal causation entropy

CausalityTools.OCE — Type

OCE <: GraphAlgorithm
OCE(; utest::IndependenceTest = SurrogateTest(MIShannon(), KSG2(k = 3, w = 3)),
      ctest::C = LocalPermutationTest(CMIShannon(), FPVP(k = 3, w = 3)),
      τmax::T = 5, α = 0.05
)

The optimal causation entropy (OCE) algorithm for causal discovery (Sun et al., 2015)[Sun2015].

Description

The OCE algorithm has three steps to determine the parents of a variable xᵢ.

Perform pairwise independence tests using utest and select the variable xⱼ(-τ) that has the highest significant (i.e. with associated p-value below α) association with xᵢ(0). Assign it to the set of selected parents P.

Perform conditional independence tests using ctest, finding the parent Pₖ that has the highest association with xᵢ given the already selected parents, and add it to P. Repeat until no more variables with significant association are found.

Backwards elimination of parents Pₖ of xᵢ(0) for which xᵢ(0) ⫫ Pₖ | P - {Pₖ}, where P is the set of parent nodes found in the previous steps.

τmax indicates the maximum lag τ between the target variable xᵢ(0) and its potential parents xⱼ(-τ).

Returns

When used with infer_graph, it returns a vector p, where p[i] are the parents for each input variable. In the future, this will return a labelled, directed graph with all the detected associations.

Examples

Inferring time series graph from a chain of logistic maps

Sun2015Sun, J., Taylor, D., & Bollt, E. M. (2015). Causal network inference by optimal causation entropy. SIAM Journal on Applied Dynamical Systems, 14(1), 73-106.