Inferring causal graphs

The API for inferring causal graphs is defined by:

infer_graph

GraphAlgorithm, and its subtypes

Causal graph API

CausalityTools.infer_graph — Function

infer_graph(algorithm::GraphAlgorithm, x) → g

Infer graph from input data x using the given algorithm.

Returns a graph g, whose type depends on algorithm.

Algorithms

CausalityTools.GraphAlgorithm — Type

GraphAlgorithm

The supertype of all algorithms used to infer causal relationships from graphs. Each GraphAlgorithm returns some graph that is compatible with the Graphs.jl package, which can be used to explore the inferred graphs further.

Concrete subtypes are:

PCRobust

PCRobust

CausalityTools.PCRobust — Type

PCRobust <: GraphAlgorithm
PCRobust(
    unconditional_test::IndependenceTest,
    conditional_test::IndependenceTest;
    α = 0.05)

The "robustified" version (Kalisch & Bühlmann, 2008[Kalisch2008]) of the PC algorithm (Spirtes et al., 2000)[Spirtes2000]. The unconditional_test must be an IndependenceTest with a valid pairwise AssociationMeasure, and conditional_test must be an IndependenceTest with a valid conditional AssociationMeasure.

Description

When used with infer_graph on some input data x, the PCRobust algorithm performs the following steps:

Initialize an empty fully connected graph g with N nodes, where N is the number of variables and x[i] is the data for the i-th node.

Reduce the fully connected g to a skeleton graph by performing pairwise independence tests between all vertices using unconditional_test. Remove any edges where adjacent vertices are found to be independent according to the test (i.e. the null hypothesis of independence cannot be rejected at significance level 1 - α).

Thin the skeleton g by conditional independence testing. If x[i] ⫫ x[j] | x[Z] for some set of variables Z (not including i and j) according to conditional_test (i.e. the null hypothesis of conditional independence cannot be rejected at significance level 1 - α), then the edge between i and j is removed, and we record the separating set S(i, j) = Z. Independence tests are first performed for conditioning sets of size 1, and repeated for conditioning sets of increasing size, which in most cases limits the number of tests needed. The separating sets S(i, j), which records which variables were in the conditioning set that rendered variables i and j independent, are recorded.

Create a directed graph dg from g by replacing every undirected edge X - Y in g by the bidirectional edge X ↔ Y (i.e. construct two directional edges X → Y and Y → X). Orientiation rules 0-3 are then repeatedly applied to dg until no more edges can be oriented:

Rule 0 (orients v-structures): X ↔ Y ↔ Z becomes X → Y ← Z if Y is not in the separating set S(X, Z).

Rule 1 (prevents new v-structures): X → Y ↔ Z becomes X → Y → Z if X and Z are not adjacent.

Rule 2 (avoids cycles): X → Y → Z ↔ X becomes X → Y → Z ← X.

Rule 3: To avoid creating cycles or new v-structures, whenever X - Y → Z, X - W → Z, and X - Z but there is no edge between Y and W, turn the undirected X - Z edge into the directed edge X → Z.

The resulting directed graph is then returned.

Info

The "PC" algorithm is named after the first names of the authors, Peter Spirtes and Clark Glymour.

Kalisch2008Kalisch, M., & Bühlmann, P. (2008). Robustification of the PC-algorithm for directed acyclic graphs. Journal of Computational and Graphical Statistics, 17(4), 773-789.

Spirtes2000Spirtes, P., Glymour, C. N., Scheines, R., & Heckerman, D. (2000). Causation, prediction, and search. MIT press.