Conditional entropy

Conditional entropy API

The mutual information API is defined by

ConditionalEntropy,

entropy_conditional,

We provide a suite of estimators of various mutual information quantities. Many more variants exist in the literature. Pull requests are welcome!

Conditional entropy definitions

CausalityTools.ConditionalEntropy — Type

The supertype for all conditional entropies.

CausalityTools.CEShannon — Type

CEShannon <: ConditionalEntropy
CEShannon(; base = 2,)

TheShannon conditional entropy measure.

Discrete definition

Sum formulation

The conditional entropy between discrete random variables $X$ and $Y$ with finite ranges $\mathcal{X}$ and $\mathcal{Y}$ is defined as

\[H^{S}(X | Y) = -\sum_{x \in \mathcal{X}, y \in \mathcal{Y}} = p(x, y) \log(p(x | y)).\]

This is the definition used when calling entropy_conditional with a ContingencyMatrix.

Two-entropies formulation

Equivalently, the following difference of entropies hold

\[H^S(X | Y) = H^S(X, Y) - H^S(Y),\]

where $H^S(\cdot$ and $H^S(\cdot | \cdot)$ are the Shannon entropy and Shannon joint entropy, respectively. This is the definition used when calling entropy_conditional with a ProbabilitiesEstimator.

Differential definition

The differential conditional Shannon entropy is analogously defined as

\[H^S(X | Y) = h^S(X, Y) - h^S(Y),\]

where $h^S(\cdot$ and $h^S(\cdot | \cdot)$ are the Shannon differential entropy and Shannon joint differential entropy, respectively. This is the definition used when calling entropy_conditional with a DifferentialEntropyEstimator.

CausalityTools.CETsallisFuruichi — Type

CETsallisFuruichi <: ConditionalEntropy
CETsallisFuruichi(; base = 2, q = 1.5)

Furuichi (2006)'s discrete Tsallis conditional entropy measure.

Definition

Furuichi's Tsallis conditional entropy between discrete random variables $X$ and $Y$ with finite ranges $\mathcal{X}$ and $\mathcal{Y}$ is defined as

\[H_q^T(X | Y) = -\sum_{x \in \mathcal{X}, y \in \mathcal{Y}} p(x, y)^q \log_q(p(x | y)),\]

when $q \neq 1$. For $q = 1$, $H_q^T(X | Y)$ reduces to the Shannon conditional entropy:

\[H_{q=1}^T(X | Y) = -\sum_{x \in \mathcal{X}, y \in \mathcal{Y}} = p(x, y) \log(p(x | y))\]

CausalityTools.CETsallisAbe — Type

CETsallisAbe <: ConditionalEntropy
CETsallisAbe(; base = 2, q = 1.5)

Abe & Rajagopal (2001)'s discrete Tsallis conditional entropy measure.

Definition

Abe & Rajagopal's Tsallis conditional entropy between discrete random variables $X$ and $Y$ with finite ranges $\mathcal{X}$ and $\mathcal{Y}$ is defined as

\[H_q^{T_A}(X | Y) = \dfrac{H_q^T(X, Y) - H_q^T(Y)}{1 + (1-q)H_q^T(Y)},\]

where $H_q^T(\cdot)$ and $H_q^T(\cdot, \cdot)$ is the Tsallis entropy and the joint Tsallis entropy.

More variants exist in the literature. Pull requests are welcome!

Discrete conditional entropy

CausalityTools.entropy_conditional — Method

entropy_conditional(measure::ConditionalEntropy, c::ContingencyMatrix{T, 2}) where T

Estimate the discrete version of the given ConditionalEntropy measure from its direct (sum) definition, using the probabilities from a pre-computed ContingencyMatrix, constructed from two input variables x and y.

The convention is to compute the entropy of the variable in the first column of c conditioned on the variable in the second column of c. To do the opposite, call this function with a new contingency matrix where the order of the variables is reversed.

If measure is not given, then the default is CEShannon().

Contingency matrix

Discrete conditional entropy can be computed directly from its sum-definition by using the probabilities from a ContingencyMatrix. This estimation method works for both numerical and categorical data, and the following ConditionalEntropy definitions are supported.

ContingencyMatrix

CEShannon	✓

CETsallisFuruichi	✓

CETsallisAbe	✓

Table of discrete conditional entropy estimators

Here, we list the ProbabilitiesEstimators that are compatible with entropy_conditional, and which definitions they are valid for.

Estimator	Principle	CEShannon	CETsallisAbe	CETsallisFuruichi

CountOccurrences	Frequencies	✓	✓	x

ValueHistogram	Binning (histogram)	✓	✓	x

SymbolicPermuation	Ordinal patterns	✓	✓	x

Dispersion	Dispersion patterns	✓	✓	x

Differential/continuous conditional entropy

Table of differential conditional entropy estimators

Continuous/differential mutual information may be estimated using any of our DifferentialEntropyEstimators that support multivariate input data.

Estimator	Principle	CEShannon	CETsallisAbe	CETsallisFuruichi

Kraskov	Nearest neighbors	✓	x	x

Zhu	Nearest neighbors	✓	x	x

ZhuSingh	Nearest neighbors	✓	x	x

Gao	Nearest neighbors	✓	x	x

Goria	Nearest neighbors	✓	x	x

Lord	Nearest neighbors	✓	x	x

LeonenkoProzantoSavani	Nearest neighbors	✓	x	x

Abe2001Abe, S., & Rajagopal, A. K. (2001). Nonadditive conditional entropy and its significance for local realism. Physica A: Statistical Mechanics and its Applications, 289(1-2), 157-164.