Independence testing

A common application of information theoretic methods such as conditional mutual information (condmutualinfo) is in the context of null hypothesis testing for the conditional independence of variables.

Depending on the context, the input data and the method used, there are many considerations to be made about how to perform this conditional independence testing. Luckily, many excellent frameworks for doing so exist in the literature.

Here, we present some commonly used independence tests from the scientific literature, which can all be seamlessly used with the function independence, with any measure that quantifies conditional independence, in combination with any compatible estimator.

For example, in just a few lines of code, you can perform Runge's local permutation (LocalPermutationTest test on your data with over 20 different estimators for the conditional mutual information. If your application rather calls for the use of traditional surrogate data, the SurrogateTest test seamlessly integrates with any time series surrogate method from TimeseriesSurrogates.jl.

Independence test API

The independence test API is defined by

independence

IndependenceTest

ConditionalIndependenceTest

CausalityTools.independence — Function

independence(test::IndependenceTest, x, y, [z]) → summary

Perform the given IndependenceTest test on data x, y and z. If only x and y are given, test must provide a bivariate association measure. If z is given too, then test must provide a conditional association measure.

Returns a test summary, whose type depends on test.

Independence tests

Missing docstring.

Missing docstring for ConditionalIndependenceTest. Check Documenter's build log for details.

Surrogate test (global permutation)

CausalityTools.SurrogateTest — Type

SurrogateTest <: IndependenceTest
SurrogateTest(measure, [est];
    nshuffles::Int = 100,
    surrogate = RandomShuffle(),
    rng = Random.default_rng(),
)

A generic (conditional) independence test for assessing whether two variables X and Y are independendent, potentially conditioned on a third variable Z, based on surrogate data. Used with independence.

Description

This is a generic one-sided hypothesis test that checks whether x and y are independent (given z, if provided) based on resampling from a null distribution assumed to represent independence between the variables. The null distribution is generated by repeatedly shuffling the input data in some way that is intended to break any dependence between the input variables.

There are different ways of shuffling, dictated by surrogate, each representing a distinct null hypothesis. For each shuffle, the provided measure is computed (using est, if relevant). This procedure is repeated nshuffles times, and a test summary is returned.

For bivariate measures, the default is to shuffle both input variables. For conditional measures accepting three input variables, the default is to shuffle the first input. Exceptions are:

If TransferEntropy measure such as TEShannon, then the source variable is always shuffled, and the target and conditional variable are left unshuffled.

Examples

Quickstart examples.

Local permutation

CausalityTools.LocalPermutationTest — Type

LocalPermutationTest <: IndependenceTest
LocalPermutationTest(measure, [est];
    kperm::Int = 5,
    nshuffles::Int = 100,
    rng = Random.default_rng())

LocalPermutationTest is a generic conditional independence test (Runge, 2018)[Runge2018] for assessing whether two variables X and Y are conditionally independendent given a third variable Z (all of which may be multivariate).

Any association measure (with a compatible estimator est, if relevant) with ordering $\hat{M}(X; Y | Z)$ (conditional variable is the third) can be used. To obtain the nearest-neighbor approach in Runge, 2018, use the CMIShannon measure with the FPVP estimator.

Description

This is a generic one-sided hypothesis test that checks whether x and y are independent (given z, if provided) based on resampling from a null distribution assumed to represent independence between the variables. The null distribution is generated by repeatedly shuffling the input data in some way that is intended to break any dependence between the input variables.

For each shuffle, the provided measure is computed (using est, if relevant) while keeping Y and Z fixed, but permuting X, i.e. $\hat{M}(\hat{X}; Y | Z)$. Each shuffle of X is done conditional on Z, such that xᵢ is replaced with xⱼ only if zᵢ ≈ zⱼ, i.e. zᵢ and zⱼ are close. Closeness is determined by a kperm-th nearest neighbor search among the points in Z, and permuted points are constructed as $(x_i^*, y_i, z_i)_{i=1}^N$, where the goal is that $x_i^*$ are drawn without replacement, and $x_i$ is replaced by $x_j$ only if $z_i \approx z_j$. This procedure is repeated nshuffles times, and a test summary is returned.

Examples

See quickstart examples.

Runge2018Runge, J. (2018, March). Conditional independence testing based on a nearest-neighbor estimator of conditional mutual information. In International Conference on Artificial Intelligence and Statistics (pp. 938-947). PMLR.