Independence testing

Independence testing API

The independence test API is defined by

independence

IndependenceTest

CausalityTools.independence — Function

independence(test::IndependenceTest, x, y, [z]) → summary

Perform the given IndependenceTest test on data x, y and z. If only x and y are given, test must provide a bivariate association measure. If z is given too, then test must provide a conditional association measure.

Returns a test summary, whose type depends on test.

Compatible independence tests

SurrogateTest.

LocalPermutationTest.

JointDistanceDistributionTest.

CausalityTools.IndependenceTest — Type

IndependenceTest <: IndependenceTest

The supertype for all independence tests.

SurrogateTest

CausalityTools.SurrogateTest — Type

SurrogateTest <: IndependenceTest
SurrogateTest(measure, [est];
    nshuffles::Int = 100,
    surrogate = RandomShuffle(),
    rng = Random.default_rng(),
)

A generic (conditional) independence test for assessing whether two variables X and Y are independendent, potentially conditioned on a third variable Z, based on surrogate data. Used with independence.

Description

This is a generic one-sided hypothesis test that checks whether x and y are independent (given z, if provided) based on resampling from a null distribution assumed to represent independence between the variables. The null distribution is generated by repeatedly shuffling the input data in some way that is intended to break any dependence between the input variables.

There are different ways of shuffling, dictated by surrogate, each representing a distinct null hypothesis. For each shuffle, the provided measure is computed (using est, if relevant). This procedure is repeated nshuffles times, and a test summary is returned. For bivariate measures, the default is to shuffle both input variables. For conditional measures accepting three input variables, the default is to shuffle the first input. Exceptions are:

If TransferEntropy measure such as TEShannon, then the source variable is always shuffled, and the target and conditional variable are left unshuffled.

Compatible measures

Measure	Pairwise	Conditional	Requires est

PearsonCorrelation	✓	✖	No

DistanceCorrelation	✓	✖	No

SMeasure	✓	✖	No

PairwiseAsymmetricInference	✓	✖	Yes

ConvergentCrossMapping	✓	✖	Yes

MIShannon	✓	✖	Yes

MIRenyiJizba	✓	✖	Yes

MIRenyiSarbu	✓	✖	Yes

MITsallisMartin	✓	✖	Yes

MITsallisFuruichi	✓	✖	Yes

PartialCorrelation	✖	✓	Yes

CMIShannon	✖	✓	Yes

CMIRenyiJizba	✖	✓	Yes

TransferEntropy	✓	✓	Yes

Examples

Pairwise test, DistanceCorrelation.

Pairwise test, TEShannon.

Conditional test, PartialCorrelation.

Pairwise test, MIShannon, categorical.

Conditional test, CMIShannon, categorical.

LocalPermutationTest

CausalityTools.LocalPermutationTest — Type

LocalPermutationTest <: IndependenceTest
LocalPermutationTest(measure, [est];
    kperm::Int = 5,
    nshuffles::Int = 100,
    rng = Random.default_rng())

LocalPermutationTest is a generic conditional independence test (Runge, 2018)[Runge2018] for assessing whether two variables X and Y are conditionally independendent given a third variable Z (all of which may be multivariate).

Any association measure (with a compatible estimator est, if relevant) with ordering $\hat{M}(X; Y | Z)$ (conditional variable is the third) can be used. To obtain the nearest-neighbor approach in Runge, 2018, use the CMIShannon measure with the FPVP estimator.

Description

This is a generic one-sided hypothesis test that checks whether x and y are independent (given z, if provided) based on resampling from a null distribution assumed to represent independence between the variables. The null distribution is generated by repeatedly shuffling the input data in some way that is intended to break any dependence between the input variables.

For each shuffle, the provided measure is computed (using est, if relevant) while keeping Y and Z fixed, but permuting X, i.e. $\hat{M}(\hat{X}; Y | Z)$. Each shuffle of X is done conditional on Z, such that xᵢ is replaced with xⱼ only if zᵢ ≈ zⱼ, i.e. zᵢ and zⱼ are close. Closeness is determined by a kperm-th nearest neighbor search among the points in Z, and permuted points are constructed as $(x_i^*, y_i, z_i)_{i=1}^N$, where the goal is that $x_i^*$ are drawn without replacement, and $x_i$ is replaced by $x_j$ only if $z_i \approx z_j$. This procedure is repeated nshuffles times, and a test summary is returned.

Examples

See quickstart examples.

JointDistanceDistributionTest

CausalityTools.JointDistanceDistributionTest — Type

JointDistanceDistributionTest <: IndependenceTest
JointDistanceDistributionTest(measure::JointDistanceDistribution; rng = Random.default_rng())

An independence test for two variables based on the JointDistanceDistribution (Amigó & Hirata, 2018)[Amigo2018].

Description

The joint distance distribution (labelled Δ in their paper) is used by Amigó & Hirata (2018) to detect directional couplings of the form $X \to Y$ or $Y \to X$. JointDistanceDistributionTest formulates their method as an independence test.

Formally, we test the hypothesis $H_0$ (the variables are independent) against $H_1$ (there is directional coupling between the variables). To do so, we use a right-sided/upper-tailed t-test to check mean of Δ is skewed towards positive value, i.e.

$H_0 := \mu(\Delta) = 0$

$H_1 := \mu(\Delta) > 0$.

When used with independence, a JDDTestResult is returned, which contains the joint distance distribution and a p-value. If you only need Δ, use jdd directly.

Examples

This example shows how the JointDistanceDistributionTest can be used in practice.

Runge2018Runge, J. (2018, March). Conditional independence testing based on a nearest-neighbor estimator of conditional mutual information. In International Conference on Artificial Intelligence and Statistics (pp. 938-947). PMLR.

Amigo2018Amigó, José M., and Yoshito Hirata. "Detecting directional couplings from multivariate flows by the joint distance distribution." Chaos: An Interdisciplinary Journal of Nonlinear Science 28.7 (2018): 075302.